add security configuration to user endpoints

+++ src/main/kotlin/dev/janssenbatista/shopping/list/api/controllers/UserController.kt
@@ -6,6 +6,7 @@ import dev.janssenbatista.shopping.list.api.services.UserService
 import jakarta.validation.Valid
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.security.core.Authentication
 import org.springframework.web.bind.annotation.*
 import java.util.*
 
@@ -20,19 +21,31 @@ class UserController(private val userService: UserService) {
     }
 
     @GetMapping("/{id}")
-    fun getUserById(@PathVariable(value = "id") userId: UUID): ResponseEntity<User> =
-            ResponseEntity.ok(userService.findUserById(userId))
+    fun getUserById(@PathVariable(value = "id") userId: UUID, authentication: Authentication): ResponseEntity<User> {
+        if (getUser(authentication).id != userId)
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+        return ResponseEntity.ok(userService.findUserById(userId))
+    }
 
     @PutMapping("/{id}")
-    fun updateUser(@PathVariable(value = "id") userId: UUID, @RequestBody @Valid userDTO: UserDTO): ResponseEntity<User> {
+    fun updateUser(@PathVariable(value = "id") userId: UUID, @RequestBody @Valid userDTO: UserDTO, authentication: Authentication): ResponseEntity<User> {
+        if (getUser(authentication).id != userId)
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         val user = User(email = userDTO.email, password = userDTO.password)
         return ResponseEntity.ok(userService.update(id = userId, user = user))
     }
 
     @DeleteMapping("/{id}")
-    fun deleteUser(@PathVariable(value = "id") userId: UUID): ResponseEntity.HeadersBuilder<*> {
+    fun deleteUser(@PathVariable(value = "id") userId: UUID, authentication: Authentication): ResponseEntity<Any> {
+        if (getUser(authentication).id != userId)
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         userService.deleteById(id = userId)
-        return ResponseEntity.noContent()
+        return ResponseEntity.noContent().build()
     }
 
+    private fun getUser(authentication: Authentication) =
+            authentication.principal as User
+
+
 }
+
